test: keep ios watchdog coverage non-spawning
